GT1-3850



查看合并请求 !132

+++ SocialSecurity/SocService/SocService-Dao/src/main/java/com/ciicsh/gto/afsupportcenter/socialsecurity/socservice/dao/SsComAccountMapper.java
@@ -80,5 +80,5 @@ public interface SsComAccountMapper extends BaseMapper<SsComAccount> {
      * @param comAccountId
      * @return
      */
-    Integer getSerialByComAccountId(@Param("comAccountId") Long comAccountId);
+    Long getSerialByComAccountId(@Param("comAccountId") Long comAccountId);
 }
